requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-core==0.1.1
langchain-huggingface==0.0.1
huggingface-hub==0.20.1
diffusers==0.25.0
transformers==4.36.2
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
Pillow==10.1.0
